Clean up field generators
